+++ ElvUI_MerathilisUI/Core/General/Changelog.lua
@@ -16,7 +16,10 @@ local DISABLED_FONT_COLOR = DISABLED_FONT_COLOR
 
 local ChangeLogData = {
 	"Changes:",
-		"• Fixed a whoopsi in Alert Code. Sorry ^o^",
+		"• Fixed an error in the ObjectiveTracker if my Shadows were disabled >.>",
+		"• Fixed an error on the Mail Frame in Classic",
+		"• Update chinese locales. Thx suyunyiu <3",
+		"• Hopefully final fix for the Alerts <.<",
 
 	" ",
 	"Notes:",