+++ ElvUI_MerathilisUI/Core/Settings/Profile.lua
@@ -820,7 +820,6 @@ MP.maps = {
 			size = E.db.general.fontSize + 2,
 			style = "OUTLINE",
 			color = {r = 1, g = 1, b = 1},
-			onlyNumber = false,
 		}
 	},
 	worldMap = {