+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Changes ###
 
-  * Fixed a whoopsi in Alert Code. Sorry ^o^
+  * Fixed an error in the ObjectiveTracker if my Shadows were disabled >.>
+  * Fixed an error on the Mail Frame in Classic
+  * Update chinese locales. Thx suyunyiu <3
+  * Hopefully final fix for the Alerts <.<